requirements.txt
PyQt5
Pillow
numpy
tensorflow